pyproject.toml
[tool.poetry]
name = "Ditto"
version = "4.0.0a"
description = "A Discord bot template for Discord.py version 2+"
authors = ["bijij <josh@josh-is.gay>"]
license = "MIT"
readme = "README.rst"
packages = [
  { include = "ditto" },
  { include = "ditto/py.typed" }
]
include = ["ditto/res/*"]

[tool.poetry.dependencies]
python = "^3.10"
donphan = ">=4.10.3"
discord-ext-menus = {git = "https://github.com/rapptz/discord-ext-menus"}
"discord.py" = ">=2.2.0"
PyYAML = ">=5.4.1"
jishaku = ">=2.5.1"
tzdata = { version = ">=2023.3", markers = "sys_platform == 'win32'" }
Pillow = ">=9.4.0"
parsedatetime = ">=2.6"
psutil = "^5.8.0"
humanize = ">=3.4.1"
aiohttp-session = "^2.9.0"
aiohttp_security = "^0.4.0"
aiohttp-jinja2 = "^1.5"

[tool.poetry.dev-dependencies]
black = ">=20.8b1"
pytest = ">=6.2.3"
isort = ">=5.10.1"

[tool.poetry-git-version-changer]
enabled = true
version-file = "ditto/__init__.py"

[tool.poetry.scripts]
start = "start:main"
test = "test:main"

[tool.black]
line-length = 125
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | res
)/
'''

[tool.isort]
profile = "black"
combine_as_imports = true
combine_star = true
line_length = 125

[tool.pytest.ini_options]
addopts = "--capture=no --assert=plain --strict-markers --tb native"
testpaths = "tests"
filterwarnings = "default"

[tool.pyright]
reportUnnecessaryTypeIgnoreComment = "warning"
pythonVersion = "3.10"
typeCheckingMode = "basic"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-git-version-changer>=1.0.4a"]
build-backend = "poetry.core.masonry.api"
